requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
ultralytics
pymongo[srv]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
fastapi_mail
boto3
razorpay
pillow
apscheduler
python-multipart
opencv-python-headless